requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
PyJWT==2.9.0
PyJWT==2.9.0
pytz==2024.1
Requests==2.32.3
